requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
arabic-reshaper==3.0.0
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==24.2.0
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
cssselect2==0.7.0
distro==1.9.0
exceptiongroup==1.2.2
fonttools==4.54.1
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.6
httpx==0.27.2
httpx-sse==0.4.0
idna==3.10
jiter==0.6.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.4
langchain-core==0.3.12
langchain-text-splitters==0.3.0
langgraph==0.2.39
langgraph-checkpoint==2.0.1
langgraph-sdk==0.1.33
langsmith==0.1.137
lxml==5.3.0
markdown2==2.5.1
msgpack==1.1.0
multidict==6.1.0
numpy==1.26.4
openai==1.52.2
orjson==3.10.10
oscrypto==1.3.0
packaging==24.1
pillow==11.0.0
propcache==0.2.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydyf==0.11.0
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
pypdf==5.0.1
pyphen==0.16.0
python-bidi==0.6.3
python-dotenv==1.0.1
PyYAML==6.0.2
qrcode==8.0
regex==2024.9.11
reportlab==4.2.5
requests==2.32.3
requests-toolbelt==1.0.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
svglib==1.5.1
tavily-python==0.5.0
tenacity==9.0.0
tiktoken==0.8.0
tinycss2==1.3.0
tqdm==4.66.5
typing_extensions==4.12.2
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.3
weasyprint==62.3
webencodings==0.5.1
xhtml2pdf==0.2.16
yarl==1.16.0
zopfli==0.2.3.post1
